refactor(admin): extract UI utilities to ui_utils module; update includes and Makefile
data_handle.h alias header added; admin and voting include it. CI and docs updated.
