+++ .github/workflows/ci.yml
@@ -2,7 +2,7 @@ name: CI
 
 on:
   push:
-    branches: [ main, cleanup/** ]
+    branches: [main, cleanup/**]
   pull_request:
 
 jobs:
