+++ REMOVAL_SUMMARY.md
@@ -4,6 +4,7 @@ Date: 2025-10-21
 Branch: cleanup/copilot
 
 This repository was simplified to keep only:
+
 - Enhanced data handling (CSV-backed CRUD) — `src/data_handler_enhanced.*`
 - Admin console (data admin logic/UI) — `src/admin.c`
 - Voting algorithm — `src/voting.*`
@@ -11,27 +12,33 @@ This repository was simplified to keep only:
 ## Removed Items
 
 ### UI / Legacy Modules
+
 - `src/main.c` (rich UI entry)
 - `src/display.c` (UI helpers)
 - `src/voter_register.c/.h`, `src/candidate_register.c/.h` (overlapping register flows)
 - `src/data_handler.c/.h` (legacy handler superseded by enhanced)
 
 ### Demos / Scripts
+
 - `demo.sh`, `demo_admin.sh`, `demo_enhanced.sh`, `demo_voting.sh`
 
 ### Documentation
+
 - `README_Enhanced.md`, `VOTING_ALGORITHM_GUIDE.md`, `VOTING_IMPLEMENTATION_COMPLETE.md`
 - `docs/ADMIN_SYSTEM_GUIDE.md`, `docs/COMPREHENSIVE_REPORT.md`, `docs/technical_doc.txt`, `docs/user_guide.txt`
 
 ### Tests (Legacy)
+
 - Root-level: `test_enhanced_data_handler.c`, `test_helpers.c`, `test.sh`
 - `test/` subfolders: debug suites and sample programs; all no longer referenced by build
 
 ## Rationale
+
 - Reduce maintenance surface area and focus on core logic
 - Avoid duplicate/conflicting implementations
 - Keep a clean, minimal build and test flow
 
 ## Notes
+
 - Data files under `data/` are retained. Admin console and tests create minimal headers if missing.
 - Legacy `test/` subfolders may remain as empty directories after file deletions; they are ignored by build.
