+++ src/data_handle.h
@@ -1,0 +1,8 @@
+#ifndef DATA_HANDLE_H
+#define DATA_HANDLE_H
+
+// Thin alias to preserve a simpler include name while keeping the
+// enhanced data handler implementation and APIs intact.
+#include "data_handler_enhanced.h"
+
+#endif // DATA_HANDLE_H
