+++ src/voting.c
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
-#include "data_handler_enhanced.h"
+#include "data_handle.h"
 #include "voting.h"
 
 // Color codes for result display
